+++ src/main/java/insane96mcp/runeenchanting/data/provider/RERuneTagProvider.java
@@ -55,7 +55,8 @@ public class RERuneTagProvider extends TagsProvider<Rune> {
                 RERunes.CURSE_OF_UNSTABLE_MOTION,
                 RERunes.CURSE_OF_STATIC_CHARGE,
                 RERunes.CURSE_OF_EXPLOSION,
-                RERunes.CURSE_OF_CORROSION);
+                RERunes.CURSE_OF_CORROSION,
+                RERunes.CURSE_OF_UNDEAD);
 
         add(VANISHABLE,
                 RERunes.CURSE_OF_VANISHING,